Move definitions of conversion functions
Move definitions of conversion functions

Move definitions of conversion functions
