+++ border-candle-agent/Cargo.toml
@@ -28,6 +28,8 @@ rand = { workspace = true }
 itertools = { workspace = true }
 ordered-float = { workspace = true }
 candle-optimisers = { workspace = true }
+num-traits = { workspace = true }
+ndarray = { workspace = true, features = ["serde"] }
 
 [dev-dependencies]
 tempdir = { workspace = true }
