+++ border/examples/gym/dqn_cartpole.rs
@@ -3,7 +3,7 @@ use border_candle_agent::{
     dqn::{Dqn, DqnConfig, DqnModelConfig},
     mlp::{Mlp, MlpConfig},
     opt::OptimizerConfig,
-    util::CriticLoss,
+    util::{arrayd_to_tensor, vec_to_tensor, CriticLoss},
     TensorBatch,
 };
 use border_core::{
@@ -17,7 +17,6 @@ use border_core::{
 };
 use border_mlflow_tracking::MlflowTrackingClient;
 use border_py_gym_env::{
-    util::{arrayd_to_tensor, vec_to_tensor},
     ArrayObsFilter, DiscreteActFilter, GymActFilter, GymEnv, GymEnvConfig, GymObsFilter,
 };
 use border_tensorboard::TensorboardRecorder;
