+++ border/examples/gym/iqn_cartpole_tch.rs
@@ -10,12 +10,12 @@ use border_core::{
 };
 use border_mlflow_tracking::MlflowTrackingClient;
 use border_py_gym_env::{
-    util::vec_to_tensor, ArrayObsFilter, DiscreteActFilter, GymActFilter, GymEnv, GymEnvConfig,
-    GymObsFilter,
+    ArrayObsFilter, DiscreteActFilter, GymActFilter, GymEnv, GymEnvConfig, GymObsFilter,
 };
 use border_tch_agent::{
     iqn::{EpsilonGreedy, Iqn as Iqn_, IqnConfig, IqnModelConfig},
     mlp::{Mlp, MlpConfig},
+    util::vec_to_tensor,
     TensorBatch,
 };
 use border_tensorboard::TensorboardRecorder;
