+++ border/examples/mujoco/sac_mujoco_async_tch.rs
@@ -14,7 +14,6 @@ use border_core::{
 use border_derive::BatchBase;
 use border_mlflow_tracking::MlflowTrackingClient;
 use border_py_gym_env::{
-    util::{arrayd_to_tensor, tensor_to_arrayd},
     ArrayObsFilter, ContinuousActFilter, GymActFilter, GymEnv, GymEnvConfig, GymObsFilter,
 };
 use border_tch_agent::{
@@ -22,6 +21,7 @@ use border_tch_agent::{
     opt::OptimizerConfig,
     sac::{ActorConfig, CriticConfig, EntCoefMode, Sac, SacConfig},
     util::CriticLoss,
+    util::{arrayd_to_tensor, tensor_to_arrayd},
     TensorBatch,
 };
 use border_tensorboard::TensorboardRecorder;
